Added mock module to requirements-tests.txt file.
As test_pipeline_files requires mock module, the file requirements-
tests.txt was created to hold modules *needed* to run the tests.

Travis' requirements already includes the mock module, but there is no
hint in the root tree that mock module is required. This new file will
help new contributors to quickly find which modules are required to run
the tests.