Try bpython if ipython isn't available in shell
Also refactor first help message printed, so that it can be passed
through to the bpython embed function's banner field. bpython usually starts
in fullscreen, and the help message is not visible when bpython is
fullscreen.

Tested ipython, bpython, and builtin shell; all print the message
visibly when first opened, and shelp() works correctly in all of them.
Also ran the automated tests, which passed, and checked shell -c, which
does not print the shelp() message, just like it shouldn't.