Add RedirectMiddleware.process_exception implementation that can follow quick redirects
Some servers, like m.huffpost.com send a 30X redirect response but close the
connection after they've assumed you've read the headers. This causes scrapy to
raise a Twisted ResponseFailed exception. We catch these exceptions in the
RedirectMiddleware, take the headers out of the response (since that much is
there), create a new scrapy Response object out of it, and return that for
processing as usual.