Improve ability to find templates
Currently template searching requires the templates to be in a
specifically configured directory (settings['TEMPLATES_DIR']).

If you change the setting for `TEMPLATES_DIR`, you no longer have
access to the standard, generic Scrapy templates.

This change first searches for a given template using the path given,
and if not found, falls back to the TEMPLATES_DIR. This allows us to
maintain project specific templates checked into version control
more easily without requiring configuration changes, or overriding the
expected available templates.

Additionally, the `.tmpl` suffix was being added regardless of what was
passed in. So if you wrote `-t my_template.tmpl` it would raise an
error. Instead, this will now check if you've already included `.tmpl`
and not add it an additional time.