mkdir -p $CACHE_DIRECTORY/argent_labrose/http_cache $STATE_DIRECTORY/argent_labrose
I thought scrapy was doing this for me, but
on moving to a new dev host, it complained.
So I guess I hotpatched this early on and did not realize?
This is bodgy, but Good Enough For Now.

PS: also force canonicalization of the paths (.resolve()), because
it is cheap and there is no reason NOT to (paranoia).