Make the build reproducible
Whilst working on the Reproducible Builds effort [0] I noticed that
scrapy could not be built reproducibly.

This is due to the documentation embedding the current build year
in the generated files, therefore making the build vary depending
on when you build it. The fix is to use SOURCE_DATE_EPOCH [1] if
it is exported to the surrounding environmnent.

I originally filed this in Debian as bug #983852 [2].

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/
 [2] https://bugs.debian.org/983852