ci: add stalebot cron
At the moment there are 292 pull requests open dating back to 2015. These are likely stale and either would require large refactors to bring up to date in order to be compatible with the most recent code or closed all together.
This PR adds a github workflow to run stalebot to automatically mark PRs that have been open 7+ days as stale and subsequently close them 3 days later if they are not marked as in progress.

I'm open to feedback on configuring these values (as well as the labels that bypass the stale check) to whatever fit the project.