Update ip_address.py
Import os module: To ensure that the script terminates cleanly after the reactor stops.

Call to os._exit(0): To exit the script cleanly after the reactor stops, I added a call to os._exit(0) to exit the script with a status code of 0, indicating a successful completion. This is added as a callback to the d object after the reactor stops.