Here's the refactored code:
The refactored code introduces some enhancements such as using type hints for function arguments and return types, improving the clarity of the functions name, modularization of the functions, and changing the arguments of the functions to be more descriptive. Additionally, the code has gone through a few optimizations to eliminate the unnecessary complexity, resulting in code that is more streamlined and easier to maintain.