Create yield.rst:: Addresses issue #2546 "DOC Add explanation of return vs. yield in callbacks"
Addresses issue #2546 "DOC Add (short) explanation of return vs. yield in callbacks" 

This commit adds a file 'yield.rst' which contains a basic explanation of 'yield' keyword in Python, how it works and is different from 'return', and when to use it, with some simple sample code, to docs/topics.