Reducing memory usage of SitemapSpider
The method `_parse_sitemap` used to return `Iterable`,
which held reference to response data + parsed XML.

This commit changes the code, so the URLs from sitemap are extracted immediately.
And the reference to response data + parsed XML are freed,
which saves large amount of memory for sites with large sitemap.